// components/types.ts
export type TextSize =
  | "type-200"
  | "type-300"
  | "type-400"
  | "type-500"
  | "type-600"
  | "type-700"
  | "type-800"
  | "type-900";

export type TextType =
  | "default"
  | "title"
  | "defaultSemiBold"
  | "subtitle"
  | "link";

export type ButtonTheme = "primary" | "secondary" | "tertiary";
export type ButtonTag = "Pressable" | "TouchableOpacity";